+++ changelog.md
@@ -38,6 +38,7 @@
 - `Client.sticker_guild_get_all` wont create a partial guild.
 - `Client.sticker_guild_get_all` now returns a list emojis.
 - Add missing `InteractionResponseContext.__repr__`.
+- Add `Message.has_any_content_field`.
 
 #### ext.slash